HOMEWORK-07:  Поправила реализацию для SQLite
